pyproject.toml
[tool.poetry]
name = "playwright_project"
version = "0.1.0"
description = ""
authors = ["ruslanmustafin <rusmus.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-dotenv = "^0.5.2"
ujson = "^5.3.0"
passlib = "^1.7.4"
pytz = "^2022.1"
allure-pytest = "^2.9.45"
pytest-xdist = "^2.5.0"
requests = "^2.27.1"
factory-boy = "^3.2.1"
bcrypt = "^3.2.2"
PyHamcrest = "^2.0.3"
pytest_check = "^1.0.5"
pytest-rerunfailures = "^10.3"

beautifulsoup4 = "^4.11.1"
filelock = "^3.8.0"
pytest-playwright = "^0.3.3"
starlette = "^0.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"